requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
galaxy.plugin.api==0.65.1
galaxyutils==0.1.5
idna==2.9
multidict==4.7.6
yarl==1.4.2
